// src/app/feature/elementos/models/elementos.ts
export interface Elemento {
  id: string;
  name :string;
  supplierId: 0;
  supplierName: string;
  elementTypeId: string;
  elementTypeName: string;
  code: string;
  elementId:string;
  elementName: string;
  unitPrice: number;
  updateDate: string;
  lastUnitPrice: number;
  currentPrice: number;
  percentage: string;
  isActive: boolean;
  distributionType: string;
}

// export interface Supplier{
//     id: string;
//     name: string;
//     nit: string;
// }